djdjfjs

+++ ProblemeDuVoyageurDeCommerce/src/population/Population.java
@@ -110,11 +110,11 @@ public class Population {
 	villesEnfant2 = indi2part1;
 	villesEnfant2.addAll(indi1part2);
         Individu enfant2 = new Individu(new Chemin(villesEnfant2, Individu.calculFitness(villesEnfant2)), Individu.calculFitness(villesEnfant2));
-            if (Math.random() <= tauxMutation){ //PAS BONNE CONDITON, AIDEZ MOI
+            if ((Double)Math.random() <= tauxMutation){ //PAS BONNE CONDITON, AIDEZ MOI
                 enfant1.mutation();
             }
             enfants.add(enfant1);
-            if (Math.random() <= tauxMutation){ //PAS BONNE CONDITON, AIDEZ MOI
+            if ((Double)Math.random() <= tauxMutation){ //PAS BONNE CONDITON, AIDEZ MOI
                enfant2.mutation();
             }
             enfants.add(enfant2);
